refactor: remove readsService, consolidate into readsController
The readsService.ts exported fetchAllReads() which was never called anywhere.
It only re-exported the ReadItem type from readsController. All imports now
point directly to readsController.
