+++ src/services/meCache.ts
@@ -1,7 +1,7 @@
 import { Highlight } from '../types/highlights'
 import { Bookmark } from '../types/bookmarks'
 import { BlogPostPreview } from './exploreService'
-import { ReadItem } from './readsService'
+import { ReadItem } from './readsController'
 
 export interface MeCache {
   highlights: Highlight[]
