+++ src/utils/linksFromBookmarks.ts
@@ -1,5 +1,5 @@
 import { Bookmark } from '../types/bookmarks'
-import { ReadItem } from '../services/readsService'
+import { ReadItem } from '../services/readsController'
 import { KINDS } from '../config/kinds'
 import { fallbackTitleFromUrl } from './readItemMerge'
 import { enhanceReadItemsWithOpenGraph } from '../services/opengraphEnhancer'
